requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
Pillow==10.2.0
pytesseract==0.3.10
qrcode==7.4.2
google-generativeai==0.3.2
python-dotenv==1.0.1
Flask-WTF==1.2.1
Werkzeug==3.0.1
gunicorn==21.2.0
requests==2.31.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1 